pyproject.toml
[tool.poetry]
name = "python-base-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "python_base_project"}]

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.24.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
fastapi = "^0.101.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.209.2"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
